Move all database integration to Databricks root
