+++ app/(chat)/api/chat/[id]/stream/route.ts
@@ -3,8 +3,8 @@ import {
   getChatById,
   getMessagesByChatId,
   getStreamIdsByChatId,
-} from '@/lib/db/queries';
-import type { Chat } from '@/lib/db/schema';
+} from '@/databricks/db/queries';
+import type { Chat } from '@/databricks/db/schema';
 import { ChatSDKError } from '@/lib/errors';
 import type { ChatMessage } from '@/lib/types';
 import { createUIMessageStream, JsonToSseTransformStream } from 'ai';
