+++ app/(chat)/api/chat/route.ts
@@ -18,8 +18,8 @@ import {
   getMessagesByChatId,
   saveChat,
   saveMessages,
-} from '@/lib/db/queries';
-import { updateChatLastContextById } from '@/lib/db/queries';
+} from '@/databricks/db/queries';
+import { updateChatLastContextById } from '@/databricks/db/queries';
 import { convertToUIMessages, generateUUID } from '@/lib/utils';
 import { generateTitleFromUserMessage } from '../../actions';
 import { myProvider } from '@/lib/ai/providers';
