+++ app/(chat)/api/vote/route.ts
@@ -1,5 +1,9 @@
 import { getAuthSession } from '@/databricks/auth/databricks-auth';
-import { getChatById, getVotesByChatId, voteMessage } from '@/lib/db/queries';
+import {
+  getChatById,
+  getVotesByChatId,
+  voteMessage,
+} from '@/databricks/db/queries';
 import { ChatSDKError } from '@/lib/errors';
 
 export async function GET(request: Request) {
