+++ artifacts/actions.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import { getSuggestionsByDocumentId } from '@/lib/db/queries';
+import { getSuggestionsByDocumentId } from '@/databricks/db/queries';
 
 export async function getSuggestions({ documentId }: { documentId: string }) {
   const suggestions = await getSuggestionsByDocumentId({ documentId });
