+++ artifacts/text/client.tsx
@@ -10,7 +10,7 @@ import {
   RedoIcon,
   UndoIcon,
 } from '@/components/icons';
-import type { Suggestion } from '@/lib/db/schema';
+import type { Suggestion } from '@/databricks/db/schema';
 import { toast } from 'sonner';
 import { getSuggestions } from '../actions';
