+++ components/sidebar-history-item.tsx
@@ -1,4 +1,4 @@
-import type { Chat } from '@/lib/db/schema';
+import type { Chat } from '@/databricks/db/schema';
 import {
   SidebarMenuAction,
   SidebarMenuButton,
