+++ components/text-editor.tsx
@@ -6,7 +6,7 @@ import { EditorState } from 'prosemirror-state';
 import { EditorView } from 'prosemirror-view';
 import React, { memo, useEffect, useRef } from 'react';
 
-import type { Suggestion } from '@/lib/db/schema';
+import type { Suggestion } from '@/databricks/db/schema';
 import {
   documentSchema,
   handleTransaction,
