+++ databricks/utils/auth-headers.ts
@@ -1,6 +1,6 @@
 import 'server-only';
 
-import { getUserFromHeaders, type User } from '@/lib/db/queries';
+import { getUserFromHeaders, type User } from '@/databricks/db/queries';
 
 export type UserType = 'guest' | 'regular';
 
@@ -18,7 +18,9 @@ export interface AuthSession {
  * Get user from Databricks Apps headers or local development environment
  * This replaces NextAuth for environments where we get user info from HTTP headers
  */
-export async function authFromHeaders(request: Request): Promise<AuthSession | null> {
+export async function authFromHeaders(
+  request: Request,
+): Promise<AuthSession | null> {
   try {
     const user: User = await getUserFromHeaders(request);
 
@@ -27,7 +29,7 @@ export async function authFromHeaders(request: Request): Promise<AuthSession | n
         id: user.id,
         email: user.email || '',
         type: 'regular' as UserType,
-      }
+      },
     };
   } catch (error) {
     console.error('[authFromHeaders] Failed to get user from headers:', error);
@@ -41,27 +43,35 @@ export async function authFromHeaders(request: Request): Promise<AuthSession | n
 export function shouldUseHeaderAuth(request: Request): boolean {
   // Use header auth if X-Forwarded-User is present (Databricks Apps)
   // or if we're in local development without NextAuth session
-  return request.headers.has('X-Forwarded-User') || process.env.NODE_ENV === 'development';
+  return (
+    request.headers.has('X-Forwarded-User') ||
+    process.env.NODE_ENV === 'development'
+  );
 }
 
 /**
  * Check if we should use header-based auth with Next.js headers object
  */
 export function shouldUseHeaderAuthFromHeaders(headersList: Headers): boolean {
-  return headersList.has('X-Forwarded-User') || process.env.NODE_ENV === 'development';
+  return (
+    headersList.has('X-Forwarded-User') ||
+    process.env.NODE_ENV === 'development'
+  );
 }
 
 /**
  * Get user from headers for Next.js page components (using headers() from 'next/headers')
  */
-export async function authFromNextHeaders(headersList: Headers): Promise<AuthSession | null> {
+export async function authFromNextHeaders(
+  headersList: Headers,
+): Promise<AuthSession | null> {
   try {
     // Create a mock request object for the existing function
     const mockRequest = {
       headers: {
         get: (name: string) => headersList.get(name),
-        has: (name: string) => headersList.has(name)
-      }
+        has: (name: string) => headersList.has(name),
+      },
     } as Request;
 
     const user = await getUserFromHeaders(mockRequest);
@@ -71,10 +81,13 @@ export async function authFromNextHeaders(headersList: Headers): Promise<AuthSes
         id: user.id,
         email: user.email || '',
         type: 'regular' as UserType,
-      }
+      },
     };
   } catch (error) {
-    console.error('[authFromNextHeaders] Failed to get user from headers:', error);
+    console.error(
+      '[authFromNextHeaders] Failed to get user from headers:',
+      error,
+    );
     return null;
   }
 }
