+++ lib/artifacts/server.ts
@@ -2,8 +2,8 @@ import { codeDocumentHandler } from '@/artifacts/code/server';
 import { sheetDocumentHandler } from '@/artifacts/sheet/server';
 import { textDocumentHandler } from '@/artifacts/text/server';
 import type { ArtifactKind } from '@/components/artifact';
-import type { Document } from '../db/schema';
-import { saveDocument } from '../db/queries';
+import type { Document } from '../../databricks/db/schema';
+import { saveDocument } from '../../databricks/db/queries';
 import type { AuthSession } from '@/databricks/auth/databricks-auth';
 import type { UIMessageStreamWriter } from 'ai';
 import type { ChatMessage } from '../types';
